BUG#28907677 - CONTRIBUTION: MY_GETNCPUS BASED ON THREADS AVAILABLE
The number of online VCPUs, as reported by sysconf, can significantly over
estimate the number of available VCPUs. E.g. When using numactl, taskset,
systemd or docket containers.

With this patch the number of available VCPUs is calculated as the number of
CPUs which can be affined to the process (using OS specific APIs). If this
information is not available it falls back to the sysconf info as before.
This patch is based on a contribution patch provided by Daniel Black.

Change-Id: I05644d3d10827f68521ad9a53786e6e04a06453d