Bug#28977444 - DD CODE DOES NOT CHECK RETURNED DD OBJECT FOR ACQUIRE_UNCACHED_UNCOMMITTED
For DD code which acquires dd::tablespace object, it will have to call
client->acquire_uncached_uncommitted() to get the object. Though the
function return value indicates whether the call succeeded of failed,
the passed out dd::tablespace pointer also tells if the DD object got
or not.

Actually, if this DD object passed out is nullptr, it means the data
dictionary is somehow corrupted, at least for current table, because
this scenario, when a table exists but the tablespace information
tied to it disappears, is surely not expected. However, it would be
nice to double check the passed out dd::tablespace object to make code
more robust and easier reading.

So the fix will clean up all calls to
client->acquire_uncached_uncommitted() in DD code and double check
the dd::tablespace object after those calls.

RB: 21065
Reviewed-by: Kevin Lewis <kevin.lewis@oracle.com>