Bug #27809371: CLIENTS DO NOT EXIT FOR INVALID --SSL-FIPS-MODE SETTINGS
Problem: On platforms where fips mode is not supported, if we start clients
on these platforms with --ssl-fips-mode=ON/STRICT then client reports an error
and continues.

Analysis: When --ssl-fips-mode=ON/STRICT cannot be set there is no return code
returned by function which sets this option. This function is
set_client_ssl_options(). Thus error code path is lost and execution further
continues.

Fix: Fix is to make set_client_ssl_options() function return an error in case
fips mode cannot be set, so that caller of this function can check for this
return code and act accordingly.