Bug#28987365 MYSQL DOESN'T START ON SOLARIS X86 OPTIMIZED
The problem is that the static thread-local variable 'tables' in the
temptable engine code is not always properly initialized before being used
(seems to be a compiler bug on this platform).

The solution is to clean up declaration/definition of 'tables',
and to use 'make_tuple' rather than 'forward_as_tuple' when using 'emplace'
for new entries in the map.

Change-Id: I90e221b1ece90e257e24d065e9a1ed8fa3cfe5bc