Bug#28424366 : ADAPT CLUSTERJ TO USE CONNECTOR/J 8.0.12 FOR TESTING
Post push fix. The previous patches for this bug adapted ClusterJ to
use the latest Connector/J for testing. The recent versions of
Connector/J require atleast Java 8 or above to run. So transitively
now ClusterJ needs Java 8 or above. This patch updates the CMakeLists
to reflect this and also updates the MTR clusterj test to look for a
minimum version of Java 8 to run the tests.

Change-Id: Ie5c21c30b3a21e92b28faec9d6dce592a9f166e4