Bug#28330922 - KEYRING MIGRATION FAILS, BUT SAYS "KEYRING                MIGRATION SUCCESSFULL" + EMPTY FILE
DESCRIPTION
===========
If you attempt to perform a keyring migration with an
operating system user that does not have access to write to
the keyring file, then the migration will fail, but the log
still says the migration succeeded.

ANALYSIS
========
keyring_init() returns FALSE (success) or TRUE (error) in
different scenarios. In the aforesaid scenario even if
keyring_init() fails, it returns FALSE which results the
caller to assume that the migration was successful.

It should return TRUE for the error cases.

FIX
===
Return TRUE instead of FALSE. Also that the typo
"successfull" is replaced with "successful".