BUG#28841543 - NODES CAN JOIN USING WHITELIST WITH HOSTNAMES, AND LOCAL AND GR SEEDS WITH IPV6
Problem
===========================

Having a 3 node scenario (A,B,C) in which:
- Local Address is IPv6
- Seeds are IPv6
- Whitelist only has hostnames

Node A boots the group and B and C are able to join. According to the
IPv6 WL implementation, all names should be resolved to IPv4, if that
is an option. As such, nodes B and C should have not joined the group.

Analysis
===========================

What happens is that, on the whitelist IPv6 implementation, when a name
was resolved, we would expand both to IPv6 and IPv4 and compare. That would
cause the IPv6 nodes to join the group.

Fix
===========================

In the whitelist, after resolving a name, if we have one IPv4 entry,
that entry will be the one to be considered. The IPv6 entries will
not be considered.