Bug #28922795 CHECKING METADATA CACHE READINESS SUFFERS FROM TIMING ISSUES
When the routing plugin starts and runs its metadata-cache destinations
they need to know when the metadata-cache is ready to call API funtions on it.
There is API function that returns that information that we use to probe for
that every 1 ms for 100 ms. If that's not enough we leave with an error.

Changed 2 things:
 - moved seting when the metadata-cache is initialized from plugin's start()
   to init()
 - extended probing time in the routing plugin from 100ms to 1s