Bug#93240 wait_timeout error not clear and not sent to client.
The log entry for a client aborted due to wait_timeout exceeded
is not clearly stated. So this creates a specific error for it.

Also from the client perspective it only sees that the server closed
the connection, not why, so the error will be sent back to
the client before closing the connection.

Notes:
This patch does not work with YaSSL, since it does not clear its
errors in report_errors() in debug build, and ERR_clear_error is
not supported in release build. This results in write will fail
on the previous non-cleared read error (from wait_timeout).

Also debug build non-patched clients will fail an assertion,
due to it only expects async returned result/error in case of COM_QUIT.

And adding protection in net_read_packet like:
  /* Socket can't be used for reading */
  if (net->error == 2 || net->error == 4)
    return packet_error;
will make some rpl tests fail.