Bug#28957660 Regression: Crash in select_lex::flatten_subqueries
This bug hits when we have an EXISTS subquery when the subquery's
join condition is always false: It is skipped for semi-join conversion,
but when it reaches the last loop of SELECT_LEX::flatten_subqueries(),
the code attempts to dereference the substitution member. However,
EXISTS subqueries do not create substitution objects, so this failed.
The solution for this problem is to skip the final transformation for
subqueries where substitution is NULL.

This patch also solves an additional problem: flatten_subqueries() can
be modified to identify three outcomes:
1. Subquery predicates that are converted into semi-join
2. Subquery predicates that are always false.
3. The remaining subqueries.

Before this patch, subquery predicates in category (2) and (3) were
treated equally. However, subquery predicates that are always false
may be simply replaced by a false predicate. To achieve this, the
member "sj_chosen" is replaced with a three-value enum
"enum_sj_selection", and proper logic to distinguish the three cases.