BUG#29028838 LOG.WRITE_AHEAD_END_OFFSET INITS TO ERROR VALUE
Fixed the initial value. The write-ahead was not working until we switched
to a next log file. There was a simple mistake in the line where the
write_ahead_end_offset was initialized, when MySQL was being started.

Instead of current_file_real_offset it was written current_file_end_offset.

The fix is a contribution of Yuhui Wang who noticed the bug.