Bug #28992826 TEST FAILURE NDB_BINLOG.NDB_BINLOG_RESTORE ON PB2
Remove file ndb_binlog_restore-master.opt to prevent failure of test
ndb_binlog_restore in Windows instead of skipping it.

Server variable "--lower_case_tables_names" when set to "0" on windows
platform which does not support case sensitive file operations leads
to problems. A warning message is printed in the error log while
starting the server with "--lower_case_tables_names=0". Also
according to the documentation, seting "lower_case_tables_names" to
"0" on a case-insensitive filesystem might lead to index corruption.

--lower_case_tables_names was being set to 0 in
ndb_binlog_restore-master.opt which should had been removed in all
versions but wasn't removed in version 7.2 and hence was causing
failures in Windows for this version. Hence, remove this file to skip
this test for platforms which do not support case sensitive file
operations instead of resulting to failures.