BUG#27850794: MTR SHOULD CHECK FORMAT OF DISABLED.DEF
Description:
------------
MTR should check the format of an entry in a disabled.def file. It should
make sure that it always contains either a BUG# or a WL#. No test should be
disabled without a BUG# or a WL#.

Also, it is not necessary to have personal names, date field in a disable
test entry, having a bug or a worklog reference should be enough. This will
make the the disabled.def format parsable by scripts.

Fix:
----
1. Introduced the following format for an entry in a disabled.def file.

   "<suite>.<test> [@platform|@!platform] : <BUG|WL>#<XXXX> ["<comment>"]"

2. Removed personal name and date field from comment section.

3. MTR will throw an error if an entry doesn't follow the above format.

4. Length of a comment section must not be more than 80 characters.

Change-Id: I32da93c67f34c5bba4cc6f690db0417f5efdead8