Bug#29058732 CONDITION PUSHDOWN INCORRECTLY REJECTED FOR TEMPORAL DATA TYPES
Depending of whether the temporal field, or the temporal string constant
was deffined first in a condition, the condition was either pushed or not.

For such condition it is the intention to omit the string data type
semantic checks (string length and character sets) - The string
is only used as a convenient way of storing the constant date/time
string. The string value is converted to the appropriate temporal
data type before the interpreter code is generated. So these
string-checks are not relevant for temporal data types.

Skipping of these string vs temporal type check was correctly
implemented for predicates:

  <temporal_field> <op> <temporal_string_const>

Not for the synonym predicate, though:

  <temporal_string_const> <inverted_op> <temporal_field>

This patch moves the 'context->expecting_max_length()' check
from the mandatory field type check, and instead make
it conditional if only:

  'if (..STRING_RESULT && !is_supported_temporal_type(type))'

Note that this is 'symetric' with how we 'expect' a STRING_ITEM
as next argument if the Field' is the first argument. In this
case we do *not* set up an expected collation or length.