Bug#28970261: Sig6 in decorrelate_equality()
Problem occurs when a component in the WHERE clause of a subquery
is non-deterministic, in this case it is because a user variable
is referenced. Referencing a user variable is not non-deterministic
per se, but should be considered so if the statement was also setting
the same user variable.

This bug fix does not attempt to fix the underlying problem of
marking non-determinism, but merely ignores the predicates that have
non-deterministic components.

We also acknowledge that non-deterministic subqueries as a general
rule should be evaluated per outer row, and thus not be target for
materialization strategy. This is prevented by not decorrelating
the non-deterministic predicates. We also enforce the same strategy
selection for non-semijoin non-deterministic subqueries (when possible).