Bug#22180480 NDB : MYSQLD SUBSCRIPTION LOGS DO NOT INCLUDE FULL BITMAP
The log printouts of ndbcluster(the plugin) only printed two words of
the bitmap which in most cases contains 8 words, this made it hard to
diagnose problems with schema distribution. Most log printouts of
bitmaps has already been fixed in similar fashion by WL#10162 and this
change takes care of the remnants.

Fix by:
- changing remaining log printouts to use the new dynamic bitmap print
function.
- removing remaining hardcoded assumptions about size of the 'slock'
column in ndb_schema and instead dynamically adapt buffers to the
size of the column. This works fine as the asynchronous handling of
events are performed serially in the ndb binlog thread. If the size of
slock column of the physical NDB table changes, the schema distribution
is released and initialized again with values matching the new physical
table. There are also already checks inplace in the schema distribution
client ensuring that the schema fulfill certain minimum requirements
before allowing to start a schema operation.
 - removing several hardcoded constants as well as the SCHEMA_SLOCK_SIZE
constant.
 - changing Ndb_schema_dist_table::get_slock_bits() to return bytes and
rename it appropriately.
 - removing the Ndb_schema_opt::slock_length variable, it was only used
in a DBUG_PRINT and the same value could also be deduced from the slock
bitmap variables.

Change-Id: I920d0993dc7b5992b4b66af1753f079526347615