BUG#29127040 ADJUST TRANSACTION HINT COUNTERS TO LESS NODES
The per node "transaction_hint" and "transaction_no_hint"
counters available in `SHOW ENGINE NDB STATUS` is removed
in order to allow adjusting to different cluster configurations.
The counters were hardcoded and being in the critical transaction
code path it's not suitable to replace them with a dynamic solution.
The counters are undocumented.

Introduce a new per session "Ndb_trans_hint_count_session" variable
which shows the number of hinted transactions started. Using
hint when starting the transaction is normally more efficient as it
tries to select a transaction coordinator close to the data. The new
variable allows comparing with "Ndb_api_trans_start_count"
to determine how many transactions was started with and without hint.
The new variable wil be visible both in GLOBAL and SESSION scope just
like all other plugin status variables, this is due to a limitation in
the plugin framework.

Also:
- remove printouts of transaction hint counters to server log file
  in Thd_ndb destructor.
- mark g_ndb_status as only used in the file where it's declared.
- fix missing include of mysql_version.h which appears when other
  includes are removed.

Add tests for "Ndb_trans_hint_count_session" showing that the
counter incrases when running some NDB transactions.

Change-Id: Ie5824fe11bb4a0e8aee871308f95f5e0a824fdc5