Bug#27135706 BINARY LOGGING IS ON BY DEFAULT
As of MySQL server 8.0.3 binary logging is enabled by default.
With this change, all MySQL servers in a MySQL cluster (possibly
several hundreds) will each write a binary log of all changes,
potentially requiring significantly more resources than before.

To compensate for this, we change the default value of ndb-log-bin
from ON to OFF. This allows binary logging of NDB tables to be
selectively enabled using --ndb-log-bin on only a few MySQL servers.

Change-Id: I28ff691c7d174684a548b7d92b2c90f58437bedf