Bug#29055263: REMOVE GR PLUGIN README FILE
Group Replication has a README file which is mostly
forgotten, the 8.0 version still mentions 5.7 for instance.
This README file was created with the assumption that the
plugin would be released stand-alone, which didn't happen.

Resolution:
The server root has a minimalist README file that points to
documentation, which covers all code.
To avoid a not-updated file to be released we will remove
the GR plugin README file.