WL#8517 Backup parallelism : Part 10 of 25
mt-backup

Add a config parameter for ndmtd named EnableMultithreadedBackup. If
set to 1, ndbmtd will attempt to execute backups as multithreaded backups.
In a multithreaded backup, all the LDMs in a node perform backup work.
The multithreaded backup directory structure is different from the existing
backup directory structure.

It is not guaranteed that a setting of EnableMultithreadedBackup=1 will
result in a multithreaded backup, since multithreaded backups are only run
if all the data nodes have more than one LDM.

If EnableMultithreadedBackup=0, backup will always be single-threaded.
The default is EnableMultithreadedBackup=1.