Bug#29133127 ASSERTION `MAYBE_NULL' FAILED |ITEM_FUNC_CONCAT::VAL_STR(STRING*)
In non-strict SQL mode, certain strings functions will return NULL to indicate
that the result is too large for the result buffer.

The concat() function has a debug-only assert that the expression is actually
nullable if we are about to return NULL.

The fix is to set maybe_null to true for concat and concat_ws if the
result string may be longer than max_allowed_packet.

Change-Id: I5605038663703801e468e8ba487100eb5584696b