Bug#28859354 PB2 FAILURE OF NDB_GLOBAL_SCHEMA_LOCK_ERROR ON SOLARIS
When MySQL Server looses connection to the cluster it will
asynchronously teardown resources. One such resource is the NDB_SHARE
for mysql.ndb_schema table which need to be acquired by the schema
distribution client when starting DDL. If the teardown hasn't completed
when starting DDL, the DDL will fail somewhat later and give a different
warning. Both these warnings are correct behavior by the product, but
since test is expecting only the first warning message it need to wait
until the teardown has completed before continuing with the test.

Fix test by waiting for the "ndb binlog thread" to complete teardown.
Teardown has completed when the threads state is set to "Waiting for
ndbcluster to start" and can be examined via the
INFORMATION_SCHEMA.PROCESSLIST table.

Change-Id: Ie07f2504cf24519a60dfc8103820070719463047