Bug #29047811 ASSERTION FAILED: CURRENT_THD->IS_ERROR(), SQL\ITEM_GEOFUNC.CC, LINE 5358
GIS code will debug-assert that if some computation fails, and an
exception is thrown by BOOST, we catch it, and translate it to a MySQL error
code, and set an error flag in the THD.

The problem is that the error handler for partition_info::set_used_partition()
ignores *all* errors and warnings. This was seemingly done to handle MyIsam
partition pruning errors.

Fix: remove the error handler, and re-record the few tests that now
get an extra conversion warning.

Change-Id: I6478489b46e3ab96a4c9681b8e3987e271d92d30