Bug #29132639: WL#12470: SIG 6 IN CONNECTJOINS|SQL/SQL_EXECUTOR.C
Fix an issue where overlapping semijoins (first match) and outer joins would
confuse the rewriting into iterators, triggering an assert and possibly give
wrong query results. We already handled the case of a semijoin overlapping
with a single outer join, but not handling an outer multi-join.

Since these query plans are nonsensical (the join optimizer should only ever
output them if you have tables that are not actually in use for anything),
we don't need to worry a lot about query performance, only correctness;
we solve it by reusing a similar solution from overlapping weedouts, where we
handle the semi-outer-joins as regular outer joins and remove the duplicates
at the very end.

Change-Id: I471e0cb4b589234649592df4c9218cfa0ee01393