WL#12686: Extend NDB condition pushdown to allow referring values from previous tables.
Currently we only allow condition pushdown of predicate terms which refer column
values from the same table as the condition is being pushed to.

This Worklog will enhance this, such that column values from tables earlier in
the query plan could also be referred from pushed conditions. The column values
from these tables has already been retrieved by the nested loop algorithms when
we generate the NdbScanFilter for the pushed conditions. Thus, we can use the
values from these columns as 'const' values when we generate the interpreter code.

Generally we now allow push down of predicate terms of the form:

<this_table>.<Field> comp_op <other_table>.<Field>

- Where 'this_table' is the table the condition is being pushed to
- Where 'other_table' is a previous table in the query plan.