WL #12788: Iterator executor analytics queries [6/7, ROLLUP]
Implement rollup in the iterator executor.

The biggest challenge with rollup is that we need to replace the grouping
items with NULLs in the rollup rows, even for non-nullable fields. In the old
executor, which was push-based, this was solved by switching the field list when
pushing the additional rows. With the current iterator design, which is pull-
based and has no explicit field list (arguably a shortcoming, but not one which
is too easily fix), we cannot easily do this. Instead, we replace the field list
with a set of Item_refs, which can then be pointed to different places when
rollup rows are to be returned.

Change-Id: Ifaa611547cac4e100bf163bfd079bfdaea2b697c