Bug #28523127: INNODB: ASSERTION FAILURE: DICT0DD.CC:5508:!ERROR
This is the second patch for this bug.  The first tried a method of
forcing the error ER_TOO_MANY_CONCURRENT_TRXS during DDL_Log cleanup
that was very inconsistent.  The test innodb.create_table would fail
sporadically with the additional test in it.

Also, there was a memory leak because the previous patch quit cleaning
up the DDL_Log records as soon as the error occurred.

This patch injects the error at specific places in a controlled manner,
And it continues the DDL_Log cleanup when the error occurs.

The DB_TOO_MANY_CONCURRENT_TRXS error can occur during a DDL
or at startup after recovery.  If it occurs at startup the function
DDL_Log::delete_by_ids is called which uses its own DD transaction.
Iif the attempt to get a free undo log fails with this error, it will
retry up to 10 times. Each new attempt uses a new transaction so that
each attempt will try to get an undo log from a different rollback
segment.  After 10 tries, it returns that error. A error message is put
to the error log on each attempt suggesting that more rollback segments
are needed.

The test is now in its own testcase file plus additional injection
points in innodb.tablespace_encrypt_7.

Approved by Bin in RB#21060