Bug#29155126 CRASH IN GET_MYSQL_TIME_FROM_STR_NO_WARN
The behavior is different on Windows and causes the server failure
there (only). There are two issues which contribute to the failure.

a) One instance of missing error checking in analyze_field_constant:
its evaluation of (*const_val)->is_null() may lead to an error and
this is not checked. In the Windows case, we do get an error. The
error leads to an empty string of *const_value when it is accessed
later, hence the failure.

b) The different behavior on Windows comes from Bug#29175262.

Fix: Add the missing error checking. Bug#29175262 is addressed
separately.

Added a repro test which is slightly modified from the repro to also
fail on Linux without the fix (courtesy of Knut).

Change-Id: I44f9cf0b870152c500efbb64ebe9b374e753562c