Bug#29179604 CONSTANT FOLDING CAN LOSE NULL INFO IN ALWAYS TRUE CASE IN NESTED EXPRESSION
Problem:

For nullable columns, if we find an expression that is always true (unless
the column is NULL),
we will fold this expression thus:

Example:

CREATE TABLE t(i TINYINT);
SELECT * FROM t WHERE condition;

condition:
   i < 200   is folded to   i IS NOT NULL

While correct on the top level, this is not null preserving in nested cases.
condition:

     0 = (i < 255)

would fold to

     0 = (i IS NOT NULL)

so in the NULL case we would get

     0 = 0

which would select a NULL row, erroneously. Looking at the expression for the
NULL case, we
should have

     0 = (NULL < 255) ->
     0 = NULL  ->
     NULL

Solution:

In nested cases like this, fold always true conditions to
IF(column IS NULL, NULL, TRUE) instead. In the example above that will yield

    0 = IF(i IS NULL, NULL, TRUE)

which retains the null information as required if i is NULL:

    0 = IF(TRUE, NULL, TRUE) ->
    0 = NULL ->
    NULL

Added repro test case.

Change-Id: Ia56bcdfddb6f527fae9ae9bd5681a085b79e3ec4