Bug#29024182 - [REGRESSION]INVOCATION OF MIGRATION FAILS WITH                KEYRING_OKV_OPEN_MODE=1
DESCRIPTION
===========
An attempt to migrate keyring using keyring_okv plugin
would result in server throwing error:

unknown variable 'keyring_okv_open_mode=1'.

ANALYSIS
========
Patch to Bug#28339014 makes keyring migration server set
the variable 'open_mode' for all keyring plugins. This var
was added in all the keyring plugins except okv, and hence
this error.

FIX
===
Introducing the same var in okv keyring plugin as done for
the other keyring plugins, so that the server can set it
for this plugin too. Please be noted that the var would
however wont be used in plugin as such.

Renamed the common variable
'keyring_file_open_mode' to 'keyring_open_mode' so that
it is relevant for all keyring plugins. Also prefixed the
keyword 'loose' to the variable <keyring>_open_mode.