Bug #28563639   DATA NODE CRASHES DURING BACKUP AFTER  CLEARING REDO LOG
The STOP_BACKUP stage of a backup drops the triggers created for
logging inserts/updates/deletes during the backup. Before dropping
the triggers, the signal queues are sync'ed so that all the queued
insert/update/deletes are processed. In the case of a backup which
aborts because of buffer overflow, the sync results in ABORT
signals being processed before the STOP_BACKUP can continue. The
abort changes the backup state to ABORT_BACKUP_ORD, which causes a
crash since the resuming STOP_BACKUP requires that the state
should be STOP_BACKUP_REQ.

This is fixed by changing the backup status update time. The
backup status is set to STOP_BACKUP_REQ after the sync completes.